🦄 refactor: split code between vue and core folders
