+++ my-test-examples/src/components/Droppable.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { PropType, ref } from "vue";
 import useDragAndDrop from "../../../src/vue/useDragAndDrop";
-import { DragEndEventData } from "../../../src/composables";
+import { DragEndEventData } from "../../../src/core";
 
 const list = ref([] as string[])
 const { droppableGroup, elements, onDragEnd, onDragStart } = defineProps({
