+++ src/core/utils/touchDevice.ts
