+++ src/index.ts
@@ -1,4 +1,4 @@
 import useDragAndDrop from "./vue/useDragAndDrop";
-import { DragEndEventData, DragStartEventData, Direction } from "./composables";
+import { DragEndEventData, DragStartEventData, Direction } from "./core";
 export { useDragAndDrop };
 export type {DragStartEventData, DragEndEventData, Direction}
